now just one screen renders with the menu bar, instead of two
