+++ App.js
@@ -1,10 +1,9 @@
 import * as React from "react";
-import Menu from "./components/Menu";
 import Main from "./main";
 
 function App() {
   return (
-    <><Main /><Menu /></>
+    <><Main /></>
   );
 }
