+++ components/Menu.js
@@ -1,25 +1,26 @@
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { NavigationContainer } from '@react-navigation/native';
-import * as React from 'react';
-import CalendarScreen from './CalendarScreen';
-import StartScreen from './StartScreen';
+// import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+// import { NavigationContainer } from '@react-navigation/native';
+// import * as React from 'react';
+// import CalendarScreen from './CalendarScreen';
+// import StartScreen from './StartScreen';
+// import NotesScreen from './NotesScreen';
 
+// const Tab = createBottomTabNavigator();
 
-const Tab = createBottomTabNavigator();
+// function MyTabs() {
+//   return (
+//     <Tab.Navigator>
+//       <Tab.Screen name="Start" component={StartScreen} />
+//       <Tab.Screen name="Calendar" component={CalendarScreen} />
+//       <Tab.Screen name="Notes" component={NotesScreen} />
+//     </Tab.Navigator>
+//   );
+// }
 
-function MyTabs() {
-  return (
-    <Tab.Navigator>
-      <Tab.Screen name="Home" component={StartScreen} />
-      <Tab.Screen name="Calendar" component={CalendarScreen} />
-    </Tab.Navigator>
-  );
-}
-
-export default function Menu() {
-  return (
-    <NavigationContainer>
-      <MyTabs />
-    </NavigationContainer>
-  );
-}
+// export default function Menu() {
+//   return (
+//     <NavigationContainer>
+//       <MyTabs />
+//     </NavigationContainer>
+//   );
+// }
